Initial Commit of theme Archetecture
Starting with the _s theme, I am going to cleave the frontend react
theme from the out of the box wordpress theme. They will both follow
similar layouts but this enforces the separation.

Also this commit is more of a "what files do I need?" sort of thing,
moving these around has caused it to break pretty significantly. But I
am going to re-write most of this anyway from scratch.